+++ src/makeapp/appmaker.py
@@ -1,6 +1,7 @@
 import logging
 import os
 import re
+from contextlib import chdir
 from datetime import date
 from pathlib import Path
 from typing import Any
@@ -12,7 +13,7 @@ from .exceptions import AppMakerException
 from .helpers.vcs import VcsHelper
 from .helpers.venvs import VenvHelper
 from .rendering import Renderer
-from .utils import PYTHON_VERSION, chdir, configure_logging, get_user_dir, read_ini
+from .utils import PYTHON_VERSION, configure_logging, get_user_dir, read_ini
 
 RE_UNKNOWN_MARKER = re.compile(r'{{ [^}]+ }}')
 BASE_PATH = os.path.dirname(__file__)